Mode auto path chooser into its own class.
