+++ src/main/java/frc/robot/Robot.java
@@ -6,8 +6,6 @@ package frc.robot;
 // WPILib Imports
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.Timer;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 // Systems
 import frc.robot.systems.ArmFSM;
 import frc.robot.systems.DriveFSMSystem;
@@ -31,15 +29,14 @@ public class Robot extends TimedRobot {
 	private DriveFSMSystem driveSystem;
 	private SpinningIntakeFSM spinningIntakeFSM;
 	private GroundMountFSM groundMountFSM;
+	private boolean isArmEnabled = true;
+	private boolean isDriveEnabled = true;
+	private boolean isIntakeEnabled = true;
 
 	// autonomus
 	private static boolean finishedDeposit = false;
 	private static int node = -1; // -1 is none, 0 is low, 1, mid, 2 is high
-	private static SendableChooser<FSMState> autoPathChooser;
-	private static SendableChooser<Integer> nodeChooser;
-	private boolean isArmEnabled = true;
-	private boolean isDriveEnabled = true;
-	private boolean isIntakeEnabled = true;
+	private AutoPathChooser autoPathChooser;
 	/**
 	 * This function that returns whether or not the robot has finished
 	 * 	depositing the object in autonomus.
@@ -65,21 +62,6 @@ public class Robot extends TimedRobot {
 		node = level;
 	}
 	/**
-	 * This function returns the chooser object for the auto path for the robot.
-	 * @return The chooser object which contains information on which auto path was selected.
-	 */
-	public static SendableChooser<FSMState> getAutoChooser() {
-		return autoPathChooser;
-	}
-
-	/**
-	 * This function returns the chooser object for the node for the auto path.
-	 * @return The chooser object which contains information on the node for the auto path.
-	 */
-	public static SendableChooser<Integer> getNodeChooser() {
-		return nodeChooser;
-	}
-	/**
 	 * This function is run when the robot is first started up and should be used for any
 	 * initialization code.
 	 */
@@ -87,22 +69,7 @@ public class Robot extends TimedRobot {
 	public void robotInit() {
 		System.out.println("robotInit");
 		input = new TeleopInput();
-		autoPathChooser = new SendableChooser<>();
-		autoPathChooser.setDefaultOption("Path 1", FSMState.P1N1);
-		autoPathChooser.addOption("Path 2", FSMState.P2N1);
-		autoPathChooser.addOption("Path 3", FSMState.P3N1);
-		autoPathChooser.addOption("Path 4", FSMState.P4N1);
-		autoPathChooser.addOption("Path 5", FSMState.P5N1);
-		autoPathChooser.addOption("Path 6", FSMState.P6N1);
-		autoPathChooser.addOption("Path 7", FSMState.P7N1);
-		SmartDashboard.putData("Auto Path", autoPathChooser);
-
-		nodeChooser = new SendableChooser<>();
-		nodeChooser.setDefaultOption("Low", 0);
-		nodeChooser.addOption("Mid", 1);
-		nodeChooser.addOption("High", 2);
-		nodeChooser.addOption("None", -1);
-		SmartDashboard.putData("Node", nodeChooser);
+		autoPathChooser = new AutoPathChooser();
 		if (isDriveEnabled) {
 			driveSystem = new DriveFSMSystem();
 		}
@@ -121,7 +88,6 @@ public class Robot extends TimedRobot {
 	@Override
 	public void autonomousInit() {
 		System.out.println("-------- Autonomous Init --------");
-		System.gc();
 		if (isArmEnabled) {
 			if (HardwareMap.isRobotGroundMount()) {
 				groundMountFSM .reset();
